# Implement WeakIdentityHashMap in avro-fastserde