Removed jboss.util.collection.WeakIdentityHashMap to meet LinkedIn Open Source Licenses requirement.

Re-run the JMH benchmark with the new WeakIdentityHashMap in Avro 1.4. Compared with creating Datum Reader/Writer for each record, latency reduction by using FastDatumReaderWriterUtil  -
> Deserialization : ~50%, from 2.4ms to 1.2ms;
> Serialization : ~20%, from 3.4ms to 2.6ms;

The performance improvement is the same as #37 

@FelixGV @gaojieliu @radai-rosenblatt @pravk03